refactor(jobs): extract commit-coordination routing into a JobsManager partial
#6: move CoordinatedJobContext, _TryCaptureCoordinatedContext, _RequireCoordinatedWriter,
_DeferSideEffects, the coordinated cron side-effect helpers, and the Log class into
JobsManager.CommitCoordination.cs. JobsManager.cs drops 1069 -> 932 lines; the coordination
concern is now a coherent 153-line unit. No behavior change (89/89 Jobs unit tests pass).
